feat: add CC Sarah pitch operator views — dashboard, follow-ups, proposals Kanban, slots, customer Pitch tab (guardian-validated)

+++ backend/database/migrations/2026_01_01_000002_create_smbs_table.php
@@ -41,7 +41,7 @@ return new class extends Migration
                     $table->json('coordinates')->nullable();
                 }
 
-                $table->tinyInteger('engagement_tier')->default(4);
+                $table->smallInteger('engagement_tier')->default(4);
                 $table->integer('engagement_score')->default(0);
                 $table->timestamp('last_email_open')->nullable();
                 $table->timestamp('last_email_click')->nullable();
